+++ SandWorm/Analytics/WaterLevel.cs
@@ -9,7 +9,7 @@ namespace SandWorm.Analytics
         {
         }
 
-        public override void GetGeometryForAnalysis(ref List<GeometryBase> outputGeometry, int waterLevel, Mesh mesh)
+        public override void GetGeometryForAnalysis(ref List<GeometryBase> outputGeometry, double waterLevel, Mesh mesh)
         {
             var bounds = mesh.GetBoundingBox(false);
             var waterPlane = new Plane(new Point3d(bounds.Max.X, bounds.Max.Y, waterLevel), new Vector3d(0, 0, 1));
